coordinate -> double

+++ src/com/vivarium/model/Coordinates.java
@@ -16,42 +16,42 @@ public class Coordinates {
         this.y=otherCoords.y;
     }
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     /**
      * @param x 
      * @param y
      */
 
-    public Coordinates(int x, int y) {
+    public Coordinates(double x, double y) {
         this.x=x;
         this.y=y;
     }
     /**
      * @param x
      */
-    public void setX(int x) {
+    public void setX(double x) {
         this.x=x;
     }
 
     /**
      * @param y
      */
-    public void setY(int y) {
+    public void setY(double y) {
         this.y=y;
     }
 
     /**
      * @return
      */
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
     /**
      * @return
      */
-    public int getY() {
+    public double getY() {
         return this.y;
     }
