+++ src/com/vivarium/view/AreaView.java
@@ -21,7 +21,7 @@ public class AreaView {
 
     // TODO Hack passer coords pour des cases -> à corriger, créer class square
     public void draw(Graphics g) {
-        g.drawImage(this.texture, area.getCoords().getX(), this.area.getCoords().getY(), null);
+        g.drawImage(this.texture, (int)area.getCoords().getX(), (int)this.area.getCoords().getY(), null);
     }
 
     /**
